[WebSocket] Test header + close frame in same packet
Modify the existing web platform test opening-handshake/005.html to send
the response header and the close frame in the same packet. This
verifies that Chrome passes the close data to JavaScript correctly in
this case.

This increases the coverage of the existing test, so there's no need to
retain the old behaviour.

Also remove the timeout expectations for this test, and make sure it
doesn't timeout on failure.

Closed: 1090210
Change-Id: I327c55d1517615511d6d71e1b464d5d3dc03b502